Move idl parser to its own module
